+++ include/GafferUSD/TypeIds.h
@@ -44,6 +44,7 @@ enum TypeId
 {
 
 	USDLayerWriterTypeId = 110225,
+	USDAttributesTypeId = 110226,
 
 	LastTypeId = 110250,