+++ Changes.md
@@ -201,6 +201,13 @@ Features
 --------
 
 - USDLayerWriter : Added a new node for baking the difference between two Gaffer scenes into a minimal USD layer on disk.
+- USDAttributes : Added a new node for setting USD's `kind` and `purpose`.
+
+Improvements
+------------
+
+- Light Editor : Added support for editing visualisation attributes of lights.
+- CameraTweaks : Added `Create` mode. This should be used in preference to `Replace` mode when the parameter may not exist.
 
 Fixes
 -----
@@ -209,6 +216,14 @@ Fixes
 - ShaderQuery :
   - Fixed error `ShaderQuery : "outPlug" is missing` when promoting a child plug of a query to a `Spreadsheet`.
   - Adding a child plug of a query to a `Spreadsheet` now uses the default name for the spreadsheet column.
+- Spreadsheet : Fixed missing `Remove` TweakPlug mode presets.
+- SceneAlgo : Fixed `attributeHistory()` results involving an `AttributeTweaks` node with `localise` enabled.
+
+API
+---
+
+- EditScopeAlgo : Added support for editing attributes.
+- TweakPlug : Added `Create` mode.
 
 0.61.12.0 (relative to 0.61.11.0)
 =========